Made rpc server extensible; modify GameMode to open differnt HUD through a if-statement
