+++ AirLib/src/api/RpcLibAutomobileServer.cpp
@@ -11,36 +11,7 @@
 #endif
 #include "api/RpcLibAutomobileServer.hpp"
 
-#include "common/Common.hpp"
-#include "controllers/VehicleCameraBase.hpp"
-STRICT_MODE_OFF
-#ifndef RPCLIB_MSGPACK
-#define RPCLIB_MSGPACK clmdep_msgpack
-#endif // !RPCLIB_MSGPACK
-//TODO: HACK: UE4 defines macro with stupid names like "check" that conflicts with msgpack library
-#undef check
-#include <rpc/server.h>
-#include "api/RpcLibAdapators.hpp"
-//TODO: HACK: UE4 defines macro with stupid names like "check" that conflicts with msgpack library
-STRICT_MODE_ON
-
-
 namespace msr { namespace airlib {
-
-    struct RpcLibAutomobileServer::impl {
-        impl(string server_address, uint16_t port)
-            : server(server_address, port)
-        {}
-
-        impl(uint16_t port)
-            : server(port)
-        {}
-
-        ~impl() {
-        }
-
-        rpc::server server;
-    };
 
     RpcLibAutomobileServer::RpcLibAutomobileServer(AutomobileSedanController* vechicle_controller, string server_address, uint16_t port)
         : vehicle_controller_(vechicle_controller)
@@ -65,22 +36,9 @@ namespace msr { namespace airlib {
     //required for pimpl
     RpcLibAutomobileServer::~RpcLibAutomobileServer()
     {
-        stop();
         vehicle_controller_ = nullptr;
     }
 
-    void RpcLibAutomobileServer::start(bool block)
-    {
-        if (block)
-            pimpl_->server.run();
-        else
-            pimpl_->server.async_run(4);   //4 threads
-    }
-
-    void RpcLibAutomobileServer::stop()
-    {
-        pimpl_->server.stop();
-    }
 
     bool RpcLibAutomobileServer::setVehicleControlSignals(float steeringAngle, float throttlePercentage, float brakePercentage)
     { 
@@ -90,12 +48,12 @@ namespace msr { namespace airlib {
 
     vector<uint8_t> RpcLibAutomobileServer::getImageFromCamera(uint8_t camera_id, VehicleCameraBase::ImageType_ type)
     {
-		auto result = vehicle_controller_->simGetImage(camera_id, type);
-		if (result.size() == 0) {
-			// rpclib has a bug with serializing empty vectors, so we return a 1 byte vector instead.
-			result.push_back(0);
-		}
-		return result;
+        auto result = vehicle_controller_->simGetImage(camera_id, type);
+        if (result.size() == 0) {
+            // rpclib has a bug with serializing empty vectors, so we return a 1 byte vector instead.
+            result.push_back(0);
+        }
+        return result;
     }
 
 }} //namespace
