+++ AirLib/src/api/RpcLibDroneClient.cpp
@@ -7,7 +7,7 @@
 #ifndef AIRLIB_NO_RPC
 //if using Unreal Build system then include precompiled header file first
 
-#include "api/RpcLibClient.hpp"
+#include "api/RpcLibDroneClient.hpp"
 
 #include "common/Common.hpp"
 #include <thread>
@@ -26,7 +26,7 @@ __pragma(warning( disable : 4239))
 
 namespace msr { namespace airlib {
 
-struct RpcLibClient::impl {
+struct RpcLibDroneClient::impl {
     impl(const string&  ip_address, uint16_t port, uint timeout_ms)
         : client(ip_address, port)
     {
@@ -39,19 +39,19 @@ struct RpcLibClient::impl {
 
 typedef msr::airlib_rpclib::RpcLibAdapators RpcLibAdapators;
 
-RpcLibClient::RpcLibClient(const string&  ip_address, uint16_t port, uint timeout_ms)
+RpcLibDroneClient::RpcLibDroneClient(const string&  ip_address, uint16_t port, uint timeout_ms)
 {
     pimpl_.reset(new impl(ip_address, port, timeout_ms));
 }
 
-RpcLibClient::~RpcLibClient()
+RpcLibDroneClient::~RpcLibDroneClient()
 {}
 
-bool RpcLibClient::ping()
+bool RpcLibDroneClient::ping()
 {
     return pimpl_->client.call("ping").as<bool>();
 }
-RpcLibClient::ConnectionState RpcLibClient::getConnectionState()
+RpcLibDroneClient::ConnectionState RpcLibDroneClient::getConnectionState()
 {
     switch (pimpl_->client.get_connection_state()) {
     case rpc::client::connection_state::connected: return ConnectionState::Connected;
@@ -62,93 +62,93 @@ RpcLibClient::ConnectionState RpcLibClient::getConnectionState()
         return ConnectionState::Unknown;
     }
 }
-bool RpcLibClient::armDisarm(bool arm)
+bool RpcLibDroneClient::armDisarm(bool arm)
 {
     return pimpl_->client.call("armDisarm", arm).as<bool>();
 }
-void RpcLibClient::setOffboardMode(bool is_set)
+void RpcLibDroneClient::setOffboardMode(bool is_set)
 {
     pimpl_->client.call("setOffboardMode", is_set);
 }
-void RpcLibClient::setSimulationMode(bool is_set)
+void RpcLibDroneClient::setSimulationMode(bool is_set)
 {
     pimpl_->client.call("setSimulationMode", is_set);
 }
-bool RpcLibClient::takeoff(float max_wait_seconds)
+bool RpcLibDroneClient::takeoff(float max_wait_seconds)
 {
     return pimpl_->client.call("takeoff", max_wait_seconds).as<bool>();
 }
-bool RpcLibClient::land(float max_wait_seconds)
+bool RpcLibDroneClient::land(float max_wait_seconds)
 {
     return pimpl_->client.call("land", max_wait_seconds).as<bool>();
 }
-bool RpcLibClient::goHome()
+bool RpcLibDroneClient::goHome()
 {
     return pimpl_->client.call("goHome").as<bool>();
 }
-void RpcLibClient::start()
+void RpcLibDroneClient::start()
 {
     pimpl_->client.call("start");
 }
-void RpcLibClient::stop()
+void RpcLibDroneClient::stop()
 {
     pimpl_->client.call("stop");
 }
 
 
-bool RpcLibClient::moveByAngle(float pitch, float roll, float z, float yaw, float duration)
+bool RpcLibDroneClient::moveByAngle(float pitch, float roll, float z, float yaw, float duration)
 {
     return pimpl_->client.call("moveByAngle", pitch, roll, z, yaw, duration).as<bool>();
 }
 
-bool RpcLibClient::moveByVelocity(float vx, float vy, float vz, float duration, DrivetrainType drivetrain, const YawMode& yaw_mode)
+bool RpcLibDroneClient::moveByVelocity(float vx, float vy, float vz, float duration, DrivetrainType drivetrain, const YawMode& yaw_mode)
 {
     return pimpl_->client.call("moveByVelocity", vx, vy, vz, duration, drivetrain, RpcLibAdapators::YawMode(yaw_mode)).as<bool>();
 }
 
-bool RpcLibClient::moveByVelocityZ(float vx, float vy, float z, float duration, DrivetrainType drivetrain, const YawMode& yaw_mode)
+bool RpcLibDroneClient::moveByVelocityZ(float vx, float vy, float z, float duration, DrivetrainType drivetrain, const YawMode& yaw_mode)
 {
     return pimpl_->client.call("moveByVelocityZ", vx, vy, z, duration, drivetrain, RpcLibAdapators::YawMode(yaw_mode)).as<bool>();
 }
 
-bool RpcLibClient::moveOnPath(const vector<Vector3r>& path, float velocity, float max_wait_seconds, DrivetrainType drivetrain, const YawMode& yaw_mode, float lookahead, float adaptive_lookahead)
+bool RpcLibDroneClient::moveOnPath(const vector<Vector3r>& path, float velocity, float max_wait_seconds, DrivetrainType drivetrain, const YawMode& yaw_mode, float lookahead, float adaptive_lookahead)
 {
     vector<RpcLibAdapators::Vector3r> conv_path;
     RpcLibAdapators::from(path, conv_path);
     return pimpl_->client.call("moveOnPath", conv_path, velocity, max_wait_seconds, drivetrain, RpcLibAdapators::YawMode(yaw_mode), lookahead, adaptive_lookahead).as<bool>();
 }
 
-bool RpcLibClient::moveToPosition(float x, float y, float z, float velocity, float max_wait_seconds, DrivetrainType drivetrain, const YawMode& yaw_mode, float lookahead, float adaptive_lookahead)
+bool RpcLibDroneClient::moveToPosition(float x, float y, float z, float velocity, float max_wait_seconds, DrivetrainType drivetrain, const YawMode& yaw_mode, float lookahead, float adaptive_lookahead)
 {
     return pimpl_->client.call("moveToPosition", x, y, z, velocity, max_wait_seconds, drivetrain, RpcLibAdapators::YawMode(yaw_mode), lookahead, adaptive_lookahead).as<bool>();
 }
 
-bool RpcLibClient::moveToZ(float z, float velocity, float max_wait_seconds, const YawMode& yaw_mode, float lookahead, float adaptive_lookahead)
+bool RpcLibDroneClient::moveToZ(float z, float velocity, float max_wait_seconds, const YawMode& yaw_mode, float lookahead, float adaptive_lookahead)
 {
     return pimpl_->client.call("moveToZ", z, velocity, max_wait_seconds, RpcLibAdapators::YawMode(yaw_mode), lookahead, adaptive_lookahead).as<bool>();
 }
 
-bool RpcLibClient::moveByManual(float vx_max, float vy_max, float z_min, float duration, DrivetrainType drivetrain, const YawMode& yaw_mode)
+bool RpcLibDroneClient::moveByManual(float vx_max, float vy_max, float z_min, float duration, DrivetrainType drivetrain, const YawMode& yaw_mode)
 {
     return pimpl_->client.call("moveByManual", vx_max, vy_max, z_min, duration, drivetrain, RpcLibAdapators::YawMode(yaw_mode)).as<bool>();
 }
 
-bool RpcLibClient::rotateToYaw(float yaw, float max_wait_seconds, float margin)
+bool RpcLibDroneClient::rotateToYaw(float yaw, float max_wait_seconds, float margin)
 {
     return pimpl_->client.call("rotateToYaw", yaw, max_wait_seconds, margin).as<bool>();
 }
 
-bool RpcLibClient::rotateByYawRate(float yaw_rate, float duration)
+bool RpcLibDroneClient::rotateByYawRate(float yaw_rate, float duration)
 {
     return pimpl_->client.call("rotateByYawRate", yaw_rate, duration).as<bool>();
 }
 
-bool RpcLibClient::hover()
+bool RpcLibDroneClient::hover()
 {
     return pimpl_->client.call("hover").as<bool>();
 }
 
-bool RpcLibClient::setSafety(SafetyEval::SafetyViolationType enable_reasons, float obs_clearance, SafetyEval::ObsAvoidanceStrategy obs_startegy,
+bool RpcLibDroneClient::setSafety(SafetyEval::SafetyViolationType enable_reasons, float obs_clearance, SafetyEval::ObsAvoidanceStrategy obs_startegy,
     float obs_avoidance_vel, const Vector3r& origin, float xy_length, float max_z, float min_z)
 {
     return pimpl_->client.call("setSafety", static_cast<uint>(enable_reasons), obs_clearance, obs_startegy,
@@ -156,11 +156,11 @@ bool RpcLibClient::setSafety(SafetyEval::SafetyViolationType enable_reasons, flo
 }
 
 //sim only
-void RpcLibClient::simSetPose(const Vector3r& position, const Quaternionr& orientation)
+void RpcLibDroneClient::simSetPose(const Vector3r& position, const Quaternionr& orientation)
 {
     pimpl_->client.call("simSetPose", RpcLibAdapators::Vector3r(position), RpcLibAdapators::Quaternionr(orientation));
 }
-vector<VehicleCameraBase::ImageResponse> RpcLibClient::simGetImages(vector<DroneControllerBase::ImageRequest> request)
+vector<VehicleCameraBase::ImageResponse> RpcLibDroneClient::simGetImages(vector<DroneControllerBase::ImageRequest> request)
 {
     const auto& response_adaptor = pimpl_->client.call("simGetImages", 
         RpcLibAdapators::ImageRequest::from(request))
@@ -168,7 +168,7 @@ vector<VehicleCameraBase::ImageResponse> RpcLibClient::simGetImages(vector<Drone
 
     return RpcLibAdapators::ImageResponse::to(response_adaptor);
 }
-vector<uint8_t> RpcLibClient::simGetImage(int camera_id, VehicleCameraBase::ImageType type)
+vector<uint8_t> RpcLibDroneClient::simGetImage(int camera_id, VehicleCameraBase::ImageType type)
 {
     vector<uint8_t> result = pimpl_->client.call("simGetImage", camera_id, type.toEnum()).as<vector<uint8_t>>();
     if (result.size() == 1) {
@@ -179,66 +179,66 @@ vector<uint8_t> RpcLibClient::simGetImage(int camera_id, VehicleCameraBase::Imag
 }
 
 //status getters
-Vector3r RpcLibClient::getPosition()
+Vector3r RpcLibDroneClient::getPosition()
 {
     return pimpl_->client.call("getPosition").as<RpcLibAdapators::Vector3r>().to();
 }
-Vector3r RpcLibClient::getVelocity()
+Vector3r RpcLibDroneClient::getVelocity()
 {
     return pimpl_->client.call("getVelocity").as<RpcLibAdapators::Vector3r>().to();
 }
-Quaternionr RpcLibClient::getOrientation()
+Quaternionr RpcLibDroneClient::getOrientation()
 {
     return pimpl_->client.call("getOrientation").as<RpcLibAdapators::Quaternionr>().to();
 }
 
-DroneControllerBase::LandedState RpcLibClient::getLandedState()
+DroneControllerBase::LandedState RpcLibDroneClient::getLandedState()
 {
     int result = pimpl_->client.call("getLandedState").as<int>();
     return static_cast<DroneControllerBase::LandedState>(result);
 }
 
-RCData RpcLibClient::getRCData()
+RCData RpcLibDroneClient::getRCData()
 {
     return pimpl_->client.call("getRCData").as<RpcLibAdapators::RCData>().to();
 }
 
-CollisionInfo RpcLibClient::getCollisionInfo()
+CollisionInfo RpcLibDroneClient::getCollisionInfo()
 {
     return pimpl_->client.call("getCollisionInfo").as<RpcLibAdapators::CollisionInfo>().to();
 }
 
-TTimePoint RpcLibClient::timestampNow()
+TTimePoint RpcLibDroneClient::timestampNow()
 {
     return pimpl_->client.call("timestampNow").as<TTimePoint>();
 }
 
-GeoPoint RpcLibClient::getHomeGeoPoint()
+GeoPoint RpcLibDroneClient::getHomeGeoPoint()
 {
     return pimpl_->client.call("getHomeGeoPoint").as<RpcLibAdapators::GeoPoint>().to();
 }
 
-GeoPoint RpcLibClient::getGpsLocation()
+GeoPoint RpcLibDroneClient::getGpsLocation()
 {
     return pimpl_->client.call("getGpsLocation").as<RpcLibAdapators::GeoPoint>().to();
 }
 
-bool RpcLibClient::isOffboardMode()
+bool RpcLibDroneClient::isOffboardMode()
 {
     return pimpl_->client.call("isOffboardMode").as<bool>();
 }
 
-bool RpcLibClient::isSimulationMode()
+bool RpcLibDroneClient::isSimulationMode()
 {
     return pimpl_->client.call("isSimulationMode").as<bool>();
 }
 
-std::string RpcLibClient::getDebugInfo()
+std::string RpcLibDroneClient::getDebugInfo()
 {
     return pimpl_->client.call("getDebugInfo").as<std::string>();
 }
 
-void RpcLibClient::confirmConnection()
+void RpcLibDroneClient::confirmConnection()
 {
     ClockBase* clock = ClockFactory::get();
 
@@ -249,7 +249,7 @@ void RpcLibClient::confirmConnection()
 
     std::cout << "Waiting for connection - " << std::flush;
     const TTimeDelta pause_time = 1;
-    while (getConnectionState() != RpcLibClient::ConnectionState::Connected)
+    while (getConnectionState() != RpcLibDroneClient::ConnectionState::Connected)
     {
         std::cout << "X" << std::flush;
         clock->sleep_for(pause_time);
