+++ DroneShell/src/main.cpp
@@ -12,7 +12,7 @@
 #include "common/common_utils/Utils.hpp"
 #include "common/common_utils/FileSystem.hpp"
 #include "common/common_utils/AsyncTasker.hpp"
-#include "api/RpcLibClient.hpp"
+#include "api/RpcLibDroneClient.hpp"
 #include "common/EarthUtils.hpp"
 #include "controllers/DroneCommon.hpp"
 #include "controllers/DroneControllerBase.hpp"
@@ -32,7 +32,7 @@ public:
     {
     }
 
-    RpcLibClient client;
+    RpcLibDroneClient client;
     AsyncTasker tasker;
 
     void sleep_for(TTimeDelta wall_clock_dt)
