+++ build.cmd
@@ -61,7 +61,13 @@ robocopy /MIR MavLinkCom\lib %MAVLINK_TARGET_LIB%
 REM //---------- all our output goes to Unreal/Plugin folder ----------
 if NOT exist Unreal\Plugins\AirSim\Source\AirLib mkdir Unreal\Plugins\AirSim\Source\AirLib
 robocopy /MIR AirLib Unreal\Plugins\AirSim\Source\AirLib  /XD temp
-copy AirLib\include\test.brush "%userprofile%\Documents\AirSim"
+
+REM //---------- set AIRSIM environment variable to the AirSim folder ----------
+setx AIRSIM %CD%
+
+REM //---------- copy the brush file for automobile to the AirSim folder ----------
+robocopy "AirLib\include\automobile" "%userprofile%\Documents\AirSim" *.brush
+
 REM //---------- done building ----------
 goto :eof
